rename client file in supabase directory to `client`
